[FIX] l10n_fr_invoice_addr: Correct print format for the Customer Address column
Version: 17.0+

Issue:
When the Customer Address column is included in the report pdf, the table format breaks.
This column is added when the `partner_id.commercial_partner_id` is different than the `partner_id` of the invoice or invoice refund.

Purpose of this PR:
To restore the format of the information table when the Customer Address column is included.
Similar to #176209

Steps to Reproduce on Runbot:
1) Install l10n_fr and switch to FR Company
2) Configure contacts a French company to have a child contact with a different address.
3) Create invoice with the child contact as the customer.
4) Print invoice and the informations table will be formatted incorrectly.

closes odoo/odoo#179203

X-original-commit: bda5540affd677a28f2e6deac99f5ea691520cf8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>