[FIX] l10n_mx: restrict account codes to coa sat
Steps to reproduce the issue:

1. Activate the Odoo Mexican Localization Reports module
2. In a Mexican company, create a new Account with 1 as code
3. Go to Trial Balance and download COA SAT (XML)
5. In the General Settings with developer mode active, Download XSD files
6. Go to Trial Balance and download COA SAT (XML) again
7. You get a UserError with an unclear message

Explanation:

The Mexican Chart of Accounts have clear rules regarding `account.account.code`.
The only way to verify those accounts is through the XSD files check, but they are not automatically downloaded and the error received with those files downloaded is not user friendly.

Fix reasoning:

Instead of regulating the code when downloading the XML, we will do so on creation or update of `account.account`.
The SAT codes are retrieved from a local file to avoid fetch requests on every module update or installation.

opw-4103681

closes odoo/odoo#178531

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>