[FIX] l10n_es_edi_tbai: the number sent in vendor bills should be the ref, not the name
For Bizkaia, when we send the invoices, the numfactura and seriefactura
should be the name of the vendor and the one of the customer.

In order to make it possible to correct it and to cancel an invoice
with the previous way of doing, we check in the original XML what the
dates sent were.

opw-4498833

closes odoo/odoo#195906

X-original-commit: 9a60b0094fe411f2792461a583af56bf7bd32ba9
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>