[FIX] discuss: fix the dimension of the push-to-talk mobile button
Before this commit, since https://github.com/odoo/odoo/pull/189657,

the mobile push-to-talk button was also affected by the aspect ratio
affecting all buttons of the call action list. This was making the
button too big and covering the whole call view.

This commit fixes this issue by introducing a dedicated selector
for the bar part of the call action list.

closes odoo/odoo#197884

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>