[FIX] account: undefined `duplicated_ref_ids` in bills list view
In b2200c8928b240d78f2f24ced39a0d45c928a666, we added a feature
to show a warning background in the vendor bills list view if
it had duplicated.

However, we had to add the field `duplicated_ref_ids` in the
xml view so that we could use it in the JS part. However, this
field will only be available in JS upon upgrading the `account`
module.

To avoid this, we now check that `duplicated_ref_ids` is defined
before accessing its `count`.

opw-none

closes odoo/odoo#201193

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>