[IMP] account: fiscal position map to many
countries like spain use the fiscal position feature to add a withholding tax to the product's domestic tax.
in order to do that, the domestic tax needs to replace itself (as was done in the old tax mapping)
The UI domain was restricting this (even though the data files do that).

When mapping taxes, the requested tax to map is returned if there is no mapping.
This seems to have been done for EU B2C fiscal positions (which should apply the same taxes if there is nothing mapped)
Now that domestic taxes are in both Domestic and EU b2c fiscal positions, this _**should**_ no longer be required.

(or at the very least, a tax should only map to itself if it already belongs to the fiscal position we are mapping to)

closes odoo/odoo#210769

Signed-off-by: Laurent Smet (las) <las@odoo.com>